refactor: split data access funtions into multiple domain files
